refactor(866): move Users, Consent and Governance jobs to their sections (G5 lane B3)
Four jobs leave `Humans.Infrastructure/Jobs`, following lane 5b-1's pattern: the job
lands in its owning section's `Contracts/` folder, `public`, taking the folder's
namespace. No registration site moved, no job id changed.

Ownership was re-measured from each body and from the rows it writes, not from the
filename, and one of the four does not belong where the lane brief guessed:

- `ProcessAccountDeletionsJob` -> `Humans.Users/Contracts/`. Enumerates on
  `IUserServiceRead`, anonymizes through `IAccountDeletionService`; `DeletionScheduledFor`
  is a User column. Job id was already `users-account-deletions`.
- `SuspendNonCompliantMembersJob` -> `Humans.Users/Contracts/`. Drives
  `INonCompliantMemberSuspension`, whose body has been in this section since lane 4b-2d.
  Job id `users-suspend-non-compliant`.
- `SyncLegalDocumentsJob` -> `Humans.Consent/Contracts/`, **not a Humans.Legal**. There is
  no such project; the pass runs behind Consent's `ILegalDocumentSyncRunner` over
  `consent_records` and the legal-document aggregate, the 4a placement inventory says
  Consent, and the job id is already `consent-legal-document-sync`.
- `TermRenewalReminderJob` -> `Humans.Governance/Contracts/`, **not Users**. Everything it
  reads and the one field it stamps (`RenewalReminderSentAt`) are Applications, reached
  through `IApplicationDecisionService`; its only Users call is a display-name stitch. Job
  id is already `governance-term-renewal-reminder`.

`public` under `Contracts/` is forced, not chosen: Shell names each concrete type at two
sites (`AddScoped<T>()` and the recurring roll-call) and HUM0034 makes any other public
type in a section assembly a build error.

The 2d shim and its body were **not** folded together. With the job in `Humans.Users`,
`INonCompliantMemberSuspension` has no consumer outside the section — the same position
lane 5b-1 reached on four leaf interfaces, which Peter answered "leave" (open item 24).
`ILegalDocumentSyncRunner` is in exactly that position too. Both are left public and the
reasoning is recorded at each site; shrinking interface surface needs Peter. What the
move does retire is the "this type must not move" note the 2d carve left on the job:
`AddOrUpdate<T>(id, ...)` is keyed on the job id and rewrites the stored type string at
every boot.

Reference changes are a net simplification of Base. One `<ProjectReference>` leaves
`Humans.Infrastructure` entirely - `Humans.Notifications.Contracts`, whose only consumer
here was `TermRenewalReminderJob`. Zero added: all three destination sections already
reference `Humans.Application`, `Humans.Infrastructure` and the leaves the jobs name.
Three package references added, `Hangfire.Core` on Users, Consent and Governance for
`[DisableConcurrentExecution]` / `IRecurringJob` - package, not project, so no cycle is
possible, and it was already arriving transitively.

Comment and doc drift the move caused is corrected alongside: three `Section.cs` remarks,
the four interface doc comments that placed their driver in Base, the Consent/Governance
contracts-leaf csproj rationales that counted the departing jobs among the Base consumers,
four `freshness:triggers` blocks that pinned the old job paths, and the Users/Consent/
Governance section docs plus `docs/sections/_Index.md`. `background-jobs.md` gains a
`src/Sections/**/Contracts/*Job.cs` trigger so its catalog stays covered as the Jobs
folder drains.

Three further leaf censuses named `SuspendNonCompliantMembersJob` as a Base consumer -
Shifts', Teams' and GoogleIntegration's. That was already false after lane 4b-2d and is
now provably so; each is re-derived from the current Base sources in the same commit,
which also clears two other stale names (`AuditViewerService`, `SystemTeamSyncJob`).

Refs nobodies-collective/Humans#866
