+++ docs/sections/_Index.md
@@ -112,7 +112,7 @@ Cross-check against [`design-rules.md` §8 (Table Ownership Map)](../architectur
 | **Ticket Tailor** | — | — | `TicketTailorService`, `StubTicketVendorService` (`Humans.TicketTailor.Services`) | — | — (the vendor adapter; owns no tables) |
 | **Users / Identity** | `UsersAdminDebugController`, `UnsubscribeController`, `LanguageController` | `AccountDeletionService`, `UserParticipationBackfillService`, `ExternalLoginService` | `UserService`, `AccountProvisioningService`, `UnsubscribeService`, `UserEmailProviderBackfillService`, *`CachingUserService`* | `UserRepository` | `AspNetUsers`, `AspNetUserClaims`, `AspNetUserLogins`, `AspNetUserTokens`, `AspNetRoles` (legacy), `AspNetUserRoles` (legacy) |
 | **Onboarding** | `OnboardingReviewController`, `OnboardingWidgetController` (`Humans.Onboarding.Controllers`), `WelcomeController` (Shell) | `OnboardingService` (`Humans.Onboarding.Services`, internal) | `OnboardingWidgetState` (`Humans.Onboarding.Services`, internal) | — | — (owns no tables; orchestrates Profiles, Consent, Teams and Governance through their service interfaces) |
-| **Human Lifecycle** | — (admin actions via `UsersAdminController`) | `HumanLifecycleService`, `NonCompliantMemberSuspension` (both `Humans.Users.Services`, internal — moved out of Base at G5 lane 4b-2d; published via `IHumanLifecycleService` / `INonCompliantMemberSuspension` on `Humans.Users.Contracts`) | — | — | — (owns no tables) |
+| **Human Lifecycle** | — (admin actions via `UsersAdminController`) | `HumanLifecycleService`, `NonCompliantMemberSuspension` (both `Humans.Users.Services`, internal — moved out of Base at G5 lane 4b-2d; published via `IHumanLifecycleService` / `INonCompliantMemberSuspension` on `Humans.Users.Contracts`; `SuspendNonCompliantMembersJob`, which drives the second, followed out of Base at G5 lane 5b-4 into `Humans.Users/Contracts/`) | — | — | — (owns no tables) |
 | **Early Entry** | `EarlyEntryRosterController` (`Humans.EarlyEntry.Controllers`, internal) | `EarlyEntryService` (`Humans.EarlyEntry.Services`, internal) | *`CachingEarlyEntryService`* (`Humans.EarlyEntry.Services`, internal) | — | — (owns no tables; fans out over every registered `IEarlyEntryProvider` — Camps, Shifts, Teams) |
 | **Cantina** | `CantinaController` (`Humans.Cantina.Controllers`) | — | `CantinaRosterService` (`Humans.Cantina.Services`) | — | — (reads Shifts via `IShiftManagementServiceRead`; owns no tables) |
 | **Dashboard** | — (rendered on Home) | `DashboardService`, `AdminDashboardService` | — | — | — |
