+++ src/Humans.Infrastructure/Humans.Infrastructure.csproj
@@ -41,6 +41,13 @@
       (memory/architecture/section-project-cycle-fix.md).
     -->
     <ProjectReference Include="..\Sections\Humans.Agent.Contracts\Humans.Agent.Contracts.csproj" />
+    <!--
+      Consent's and Governance's leaves both survive the G5 jobs move for SendReConsentReminderJob
+      and HumansMetricsService; SyncLegalDocumentsJob, SuspendNonCompliantMembersJob and
+      TermRenewalReminderJob left for their sections at lane 5b-4
+      (nobodies-collective/Humans#866), and Humans.Notifications.Contracts came out entirely in
+      the same move — TermRenewalReminderJob was its only consumer here.
+    -->
     <ProjectReference Include="..\Sections\Humans.Consent.Contracts\Humans.Consent.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Email.Contracts\Humans.Email.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Expenses.Contracts\Humans.Expenses.Contracts.csproj" />
@@ -54,7 +61,6 @@
     <ProjectReference Include="..\Sections\Humans.Governance.Contracts\Humans.Governance.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Issues.Contracts\Humans.Issues.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Mailer.Contracts\Humans.Mailer.Contracts.csproj" />
-    <ProjectReference Include="..\Sections\Humans.Notifications.Contracts\Humans.Notifications.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Onboarding.Contracts\Humans.Onboarding.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Surveys.Contracts\Humans.Surveys.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Teams.Contracts\Humans.Teams.Contracts.csproj" />
