+++ src/Humans.Web/Extensions/InfrastructureServiceCollectionExtensions.cs
@@ -8,7 +8,9 @@ using Humans.Application.Interfaces.Caching;
 using Humans.Application.Interfaces.GoogleIntegration;
 using Humans.Application.Interfaces.Repositories;
 using Humans.Budget.Contracts;
+using Humans.Consent.Contracts;
 using Humans.Gate.Contracts;
+using Humans.Governance.Contracts;
 using Humans.Tickets.Contracts;
 using Humans.Infrastructure.Caching;
 using Humans.Infrastructure.Configuration;
