+++ src/Humans.Web/Extensions/RecurringJobExtensions.cs
@@ -4,14 +4,17 @@ using Humans.Application.Configuration;
 using Humans.Application.Interfaces;
 using Humans.Application.Interfaces.GoogleIntegration;
 using Humans.Budget.Contracts;
+using Humans.Consent.Contracts;
 using Humans.Email.Contracts;
 using Humans.Gate.Contracts;
 using Humans.GoogleIntegration.Contracts;
+using Humans.Governance.Contracts;
 using Humans.Infrastructure.Jobs;
 using Humans.Issues.Contracts;
 using Humans.Monitor.Contracts;
 using Humans.Notifications.Contracts;
 using Humans.Tickets.Contracts;
+using Humans.Users.Contracts;
 
 namespace Humans.Web.Extensions;
