+++ src/Sections/Humans.Governance/Contracts/TermRenewalReminderJob.cs
@@ -4,13 +4,12 @@ using NodaTime;
 using Humans.Application.Extensions;
 using Humans.Application.Interfaces;
 using Humans.Email.Contracts;
-using Humans.Governance.Contracts;
 using Humans.Notifications.Contracts;
 using Humans.Application.Interfaces.Users;
 using Humans.Domain.Enums;
 using Humans.Users.Contracts;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Governance.Contracts;
 
 /// <summary>
 /// Nightly job that reminds approved members whose Colaborador / Asociado
@@ -22,6 +21,12 @@ namespace Humans.Infrastructure.Jobs;
 /// info via <see cref="IUserService"/>, so the job never touches
 /// a section DbContext directly
 /// (design-rules §2c).
+///
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-4
+/// (nobodies-collective/Humans#866) — Governance, not Users: the rows it reads and stamps
+/// are Applications, and its job id already said <c>governance-term-renewal-reminder</c>.
+/// It sits under <c>Contracts/</c> because Shell names the concrete type at registration and
+/// HUM0034 makes every other public type in a section assembly an error.
 /// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class TermRenewalReminderJob(
