+++ src/Sections/Humans.Governance/Docs/Governance.md
@@ -151,7 +151,7 @@ Three controllers serve this section directly. `BoardController` composes Govern
 - When an application is approved: the human's tier is updated on their profile (`IProfileService.SetMembershipTierAsync`), they are added to the Colaboradors or Asociados system team via `ISystemTeamSync`, an audit-log entry is written (`AuditAction.TierApplicationApproved`), an approval email is sent (`IEmailMessageFactory.ApplicationApproved` via `IEmailService.SendAsync`), and an in-app notification is dispatched (`NotificationSource.ApplicationApproved`). Email + notification are best-effort.
 - When an application is rejected: an audit-log entry is written (`AuditAction.TierApplicationRejected`), a rejection email is sent (`IEmailMessageFactory.ApplicationRejected` via `IEmailService.SendAsync`), and an in-app notification is dispatched (`NotificationSource.ApplicationRejected`). Email + notification are best-effort.
 - When an application is approved or rejected: all Board vote records for that application are deleted (atomic inside `IApplicationRepository.FinalizeAsync`).
-- A renewal reminder email + in-app notification is dispatched 90 days before term expiry (`TermRenewalReminderJob`, `NotificationSource.TermRenewalReminder`).
+- A renewal reminder email + in-app notification is dispatched 90 days before term expiry (`TermRenewalReminderJob`, `NotificationSource.TermRenewalReminder`). The job is `Humans.Governance/Contracts/TermRenewalReminderJob.cs` since G5 lane 5b-4 (nobodies-collective/Humans#866) — `public` because Shell names the concrete type when it registers and schedules it; it reads and stamps Applications only through `IApplicationDecisionService`.
 - On term expiry without renewal: the next `SystemTeamSyncJob` removes the human from the Colaboradors / Asociados system team (driven by `HasActiveApprovedTierAsync`) and calls `IUserService.DowngradeMembershipTierForExpiredAsync`, which **resets the profile's `MembershipTier`** — to another still-active tier the human holds, otherwise to `Volunteer` — and writes an `AuditAction.TierDowngraded` entry per downgrade.
 - After every write, `ApplicationDecisionService` invalidates `INavBadgeCacheInvalidator` and `INotificationMeterCacheInvalidator`; on approve/reject it also invalidates each affected voter's `IVotingBadgeCacheInvalidator` entry; on Board-vote upsert it invalidates the voter's `IVotingBadgeCacheInvalidator` entry.
 - When an account merge accepts, `AccountMergeService.AcceptAsync` fans out to all `IUserMerge` implementations; `ApplicationDecisionService.ReassignAsync` re-FKs `Application.UserId` (the applicant) from source to target. `BoardVote.BoardMemberUserId` is not re-FK'd — votes are transient, deleted on finalization.
@@ -183,5 +183,5 @@ Three controllers serve this section directly. `BoardController` composes Govern
 
 ### Touch-and-clean guidance
 
-- `OnboardingService`, `SendBoardDailyDigestJob`, `SendAdminDailyDigestJob`, `TermRenewalReminderJob`, `SystemTeamSyncJob`, and `NotificationMeterProvider` still read governance-owned tables directly for dashboards and batch jobs. Those uses are grandfathered until the sections owning those services migrate to call `IApplicationDecisionService` / `IApplicationRepository` instead.
+- `OnboardingService`, `SendBoardDailyDigestJob`, `SendAdminDailyDigestJob`, `SystemTeamSyncJob`, and `NotificationMeterProvider` still read governance-owned tables directly for dashboards and batch jobs. Those uses are grandfathered until the sections owning those services migrate to call `IApplicationDecisionService` / `IApplicationRepository` instead.
 - After nobodies-collective#584 the four board-voting methods lost their `OnboardingService` delegating wrappers and are consumed directly. Three of them — `GetBoardVotingDashboardAsync`, `GetBoardVotingDetailAsync`, `CastBoardVoteAsync` — came in with the section at G5 and are now internal, consumed only by the section's own `GovernanceBoardVotingController` at `/Governance/BoardVoting`. `GetUnvotedApplicationCountAsync` is the one that stayed cross-section, and it sits on `IApplicationServiceRead` (Shell's `NavBadgesViewComponent` and `AdminNavTree`, plus Notifications' `NotificationMeterProvider`). `CastBoardVoteAsync` returns `ApplicationDecisionResult` (not `OnboardingResult`); error keys `NotFound` and `NotSubmitted` are still returned by the service. The controller switch handles `NotFound` explicitly and maps all other error keys (including `NotSubmitted`) to the same "not votable" message via the default arm.
