+++ src/Sections/Humans.Governance/Humans.Governance.csproj
@@ -44,6 +44,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <!-- Contracts/TermRenewalReminderJob.cs — [DisableConcurrentExecution] and IRecurringJob,
+         since G5 lane 5b-4. Package, not project, so no cycle is possible; it was already
+         arriving transitively. -->
+    <PackageReference Include="Hangfire.Core" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" />
     <PackageReference Include="NodaTime" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" />
