+++ src/Sections/Humans.Teams.Contracts/Humans.Teams.Contracts.csproj
@@ -6,8 +6,10 @@
     the Shifts services, GoogleIntegration's TeamResourceService / GoogleAdminService /
     GoogleGroupSyncService / GoogleWorkspaceSyncService, AccountDeletionService and
     DuplicateAccountService all read through ITeamServiceRead or ITeamService, and
-    HumansMetricsService, SystemTeamSyncJob and SuspendNonCompliantMembersJob do the same
-    from Humans.Infrastructure. A Contracts/ *folder*
+    HumansMetricsService does the same from Humans.Infrastructure — re-derived at G5 lane
+    5b-4 (nobodies-collective/Humans#866): SystemTeamSyncJob lives in Humans.Teams itself,
+    and SuspendNonCompliantMembersJob's team reads are now NonCompliantMemberSuspension's
+    inside Humans.Users. A Contracts/ *folder*
     inside Humans.Teams would therefore make Base reference a section and cycle
     (design §15.5b, memory/architecture/section-project-cycle-fix.md).
