+++ src/Sections/Humans.Users.Contracts/INonCompliantMemberSuspension.cs
@@ -8,16 +8,17 @@ namespace Humans.Users.Contracts;
 /// resource removal, audit entry, claims/authorization cache eviction).
 /// </summary>
 /// <remarks>
-/// Carved out of <c>Humans.Infrastructure.Jobs.SuspendNonCompliantMembersJob</c>
-/// at G5 lane 4b-2d. The job class itself stays in <c>Humans.Infrastructure</c>
-/// and is now a shim: Hangfire serializes the *declaring type name* of a
-/// recurring job target, so <c>RecurringJob.AddOrUpdate&lt;SuspendNonCompliantMembersJob&gt;</c>
-/// pins that type to its assembly and namespace — a job queued or retry-delayed
-/// before a deploy that moved it could not resolve its target afterwards, and
-/// nothing in the build or the test suite would say so (batch #4 finding 31).
-/// The implementation is resolved from DI at execution time, so it is free to
-/// live inside the section that owns the membership lifecycle
+/// Carved out of <c>SuspendNonCompliantMembersJob</c> at G5 lane 4b-2d, when the job class
+/// was believed to be pinned to <c>Humans.Infrastructure</c> by Hangfire. That premise was
+/// re-measured and found false — <c>RecurringJob.AddOrUpdate&lt;T&gt;(id, …)</c> is keyed on
+/// the job id and rewrites the stored type string at every boot — so the job followed its
+/// body into this section at G5 lane 5b-4 (nobodies-collective/Humans#866) and now sits at
+/// <c>Humans.Users/Contracts/SuspendNonCompliantMembersJob.cs</c>
 /// (Peter, 2026-08-14: <c>SuspendNonCompliantMembersJob</c> → Users).
+///
+/// With both halves in one assembly this interface has no consumer outside the section. It
+/// is kept rather than folded into the job, matching lane 5b-1's open item 24 on the four
+/// leaf interfaces its jobs left behind: shrinking interface surface needs Peter.
 ///
 /// The contract is "do the thing", never "give me the rows" (design §15 step 6b):
 /// the old job body reached across seven sections' contracts from Base.
