+++ src/Sections/Humans.Users/Contracts/SuspendNonCompliantMembersJob.cs
@@ -2,27 +2,26 @@ using Hangfire;
 using Microsoft.Extensions.Logging;
 using NodaTime;
 using Humans.Application.Interfaces;
-using Humans.Users.Contracts;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Users.Contracts;
 
 /// <summary>
 /// Background job that suspends members who haven't re-consented to required documents
 /// after the grace period has expired.
 /// </summary>
 /// <remarks>
-/// A shim. The body moved into <c>Humans.Users</c> at G5 lane 4b-2d (Peter, 2026-08-14:
-/// membership lifecycle is Users, not Governance) and is reached through
-/// <see cref="INonCompliantMemberSuspension"/>.
+/// The schedule, the metric and the failure boundary; the sweep itself is
+/// <see cref="INonCompliantMemberSuspension"/>, carved into this section at G5 lane 4b-2d
+/// (Peter, 2026-08-14: membership lifecycle is Users, not Governance). Both halves are now
+/// in the same assembly — the interface is kept rather than folded in, matching lane 5b-1's
+/// call on the leaf interfaces its jobs left behind (open item 24: leave them public).
 ///
-/// <b>This type must not move, be renamed, or change namespace.</b> Hangfire serializes the
-/// declaring type name of a recurring job target, and
-/// <c>RecurringJob.AddOrUpdate&lt;SuspendNonCompliantMembersJob&gt;</c> in
-/// <c>RecurringJobExtensions</c> pins it to this assembly and namespace. A job enqueued or
-/// retry-delayed before a deploy that relocated it would fail to resolve its target
-/// afterwards, and neither the build nor the test suite would report it
-/// (G5 batch #4 finding 31). Carving the body out is safe by construction: the serialized
-/// type never moves and the implementation is resolved from DI at execution time.
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-4
+/// (nobodies-collective/Humans#866), which retired the "this type must not move" note the
+/// carve left here: <c>RecurringJob.AddOrUpdate&lt;T&gt;(id, …)</c> is keyed on the job
+/// <em>id</em> and rewrites the stored type string at every boot, so the type is free to
+/// change assembly. It sits under <c>Contracts/</c> because Shell names the concrete type
+/// at registration and HUM0034 makes every other public type in a section assembly an error.
 /// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class SuspendNonCompliantMembersJob(
