+++ src/Sections/Humans.Users/Docs/Users.md
@@ -31,7 +31,7 @@ The User aggregate and its identity surface. Profile-adjacent User properties (G
 - **Account Provisioning** creates new `User` rows from an OAuth login (`ExternalLoginService.CompleteExternalLoginAsync`, dispatched from `AccountController.ExternalLoginCallback`), a magic-link signup (`AccountController.CompleteSignup`), or an import (`AccountProvisioningService.FindOrCreateUserByEmailAsync` for ticket / MailerLite contacts). All three paths look up an existing user across `UserEmail` records and `User.Email` (with gmail/googlemail equivalence) before creating a new row.
 - **Unsubscribe** is the one-click email opt-out surface (`/Unsubscribe/{token}`) that updates the user's per-category `CommunicationPreference` via Profile's `ICommunicationPreferenceService`. New category-aware tokens redirect to the comms-preferences page; legacy campaign-only tokens (`CampaignUnsubscribe` Data Protection purpose) show the confirmation page and are treated as `MessageCategory.Marketing`. RFC 8058 one-click POST (`/Unsubscribe/OneClick`) also routes through the same service. No login required.
 - **Event Participation** is a per-user, per-year record (`Ticketed`, `Attended`, `NotAttending`, `NoShow`) derived from ticket sync, user self-declaration, and admin backfill. Owned by Users because the participation key is User + Year, not Ticket or Shift.
-- **Account deletion** is a 30-day grace period: `User.DeletionRequestedAt` + `DeletionScheduledFor` are stamped when a user requests deletion (with optional `DeletionEligibleAfter` for ticket-holder event holds). `ProcessAccountDeletionsJob` runs daily and calls `IUserService.AnonymizeExpiredAccountAsync` for each due user.
+- **Account deletion** is a 30-day grace period: `User.DeletionRequestedAt` + `DeletionScheduledFor` are stamped when a user requests deletion (with optional `DeletionEligibleAfter` for ticket-holder event holds). `ProcessAccountDeletionsJob` runs daily and calls `IUserService.AnonymizeExpiredAccountAsync` for each due user. That job and `SuspendNonCompliantMembersJob` are `Humans.Users/Contracts/*Job.cs` since G5 lane 5b-4 (nobodies-collective/Humans#866) — `public` because Shell names each concrete type when it registers and schedules it; the registration sites stay in Shell.
 - Identity sub-tables (renamed to `users`, `user_claims`, `user_logins`, `user_tokens`, `roles`, `user_roles` per Postgres convention in `UsersDbContext.OnModelCreating`) are managed by ASP.NET Identity's `UserManager<User>` / `SignInManager<User>`. Controllers may inject those framework services directly (design-rules §2a exception).
 
 ## Data Model
