+++ src/Sections/Humans.Users/Humans.Users.csproj
@@ -140,6 +140,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <!-- Contracts/ProcessAccountDeletionsJob.cs and Contracts/SuspendNonCompliantMembersJob.cs —
+         [DisableConcurrentExecution] and IRecurringJob, since G5 lane 5b-4. Package, not
+         project, so no cycle is possible; it was already arriving transitively. -->
+    <PackageReference Include="Hangfire.Core" />
     <!-- Helpers/ProfilePictureProcessor.cs — profile-picture resize/orient on upload. -->
     <PackageReference Include="Magick.NET-Q8-AnyCPU" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" />
