+++ tests/Humans.Application.Tests/Jobs/ProcessAccountDeletionsJobTests.cs
@@ -8,7 +8,6 @@ using Humans.AuditLog.Contracts;
 using Humans.Email.Contracts;
 using Humans.Application.Interfaces.Users;
 using Humans.Domain.Enums;
-using Humans.Infrastructure.Jobs;
 using Humans.Infrastructure.Services;
 using Humans.Users.Contracts;
